Simplify Wall Representation
This commit significantly simplifies the representation of wall states within the game by removing the separate bitboards for each color in favor of a single bitboard per player. Previously, the game maintained individual bitboards for each color for each player's wall, as well as an additional "wall occupancy" bitboard that indicated any tiles present, regardless of color. This structure was redundant and complex, especially when updating and checking the game state.